# api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateSchemaBookmarkRequest": {
        "properties": {
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "versionId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "versionId"
        ],
        "type": "object"
      },
...
      "CreateSchemaRequest": {
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "DuplicateRoleRequest": {
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "ModelDto": {
        "properties": {
          "archivedAt": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "fieldCount": {
            "format": "int64",
            "type": "integer"
          },
          "fileName": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "inputSchema": {
            "additionalProperties": { },
            "type": "object"
          },
          "name": {
            "type": "string"
          },
          "reportCount": {
            "format": "int64",
            "type": "integer"
          },
          "specificType": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedByAvatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedByEmail": {
            "type": "string"
          },
          "updatedByName": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "archivedAt",
          "createdAt",
          "fieldCount",
          "fileName",
          "id",
          "inputSchema",
          "name",
          "reportCount",
          "specificType",
          "type",
          "updatedAt",
          "updatedByAvatarUrl",
          "updatedByEmail",
          "updatedByName",
          "version"
        ],
        "type": "object"
      },
      "MoveSchemaBookmarkRequest": {
        "properties": {
          "versionId": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "versionId"
        ],
        "type": "object"
      },
...
      "PageDtoSchemaCatalogItemDto": {
        "properties": {
          "hasNext": {
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SchemaCatalogItemDto"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "totalItems": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "hasNext",
          "items",
          "page",
          "size",
          "totalItems"
        ],
        "type": "object"
      },
      "PermissionDto": {
        "properties": {
          "dangerous": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "key": {
            "enum": [
              "VIEW_WORKSPACE",
              "VIEW_ORGANIZATION",
              "EDIT_ORGANIZATION",
              "DELETE_ORGANIZATION",
              "TRANSFER_OWNERSHIP",
              "VIEW_MEMBERS",
              "INVITE_MEMBERS",
              "MANAGE_MEMBER_ROLES",
              "REMOVE_MEMBERS",
              "VIEW_INVITATIONS",
              "MANAGE_INVITATIONS",
              "VIEW_MODELS",
              "CREATE_MODELS",
              "EDIT_MODELS",
              "DELETE_MODELS",
              "RUN_PREDICTIONS",
              "EXPORT_PREDICTIONS",
              "REVIEW",
              "MANAGE_REVIEWS",
              "VIEW_PLUGINS",
              "MANAGE_PLUGINS",
              "VIEW_AUDIT_LOG"
            ],
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "dangerous",
          "description",
          "key",
          "label"
        ],
        "type": "object"
      },
...
      "RoleActionsDto": {
        "properties": {
          "canAssign": {
            "type": "boolean"
          },
          "canDelete": {
            "type": "boolean"
          },
          "canDuplicate": {
            "type": "boolean"
          },
          "canEdit": {
            "type": "boolean"
          },
          "canView": {
            "type": "boolean"
          }
        },
        "required": [
          "canAssign",
          "canDelete",
          "canDuplicate",
          "canEdit",
          "canView"
        ],
        "type": "object"
      },
      "RoleDefinitionDto": {
        "properties": {
          "actions": {
            "$ref": "#/components/schemas/RoleActionsDto"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "locked": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "items": {
              "$ref": "#/components/schemas/PermissionDto"
            },
            "type": "array"
          },
          "scope": {
            "enum": [
              "SYSTEM",
              "ORGANIZATION"
            ],
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "systemKey": {
            "type": [
              "string",
              "null"
            ]
          },
          "userCount": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "actions",
          "description",
          "id",
          "locked",
          "name",
          "permissions",
          "scope",
          "slug",
          "systemKey",
          "userCount"
        ],
        "type": "object"
      },
...
      "SchemaBookmarkDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "schemaId": {
            "format": "int64",
            "type": "integer"
          },
          "schemaName": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "format": "int32",
            "type": "integer"
          },
          "versionId": {
            "format": "int64",
            "type": "integer"
          },
          "versionName": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "createdAt",
          "id",
          "name",
          "schemaId",
          "schemaName",
          "updatedAt",
          "version",
          "versionId",
          "versionName"
        ],
        "type": "object"
      },
      "SchemaCatalogItemDto": {
        "properties": {
          "archivedAt": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "fieldCount": {
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "modelCount": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "organizationId": {
            "format": "int64",
            "type": "integer"
          },
          "reportCount": {
            "format": "int64",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedByAvatarUrl": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedByEmail": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedByName": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "archivedAt",
          "createdAt",
          "description",
          "fieldCount",
          "id",
          "modelCount",
          "name",
          "organizationId",
          "reportCount",
          "updatedAt",
          "updatedByAvatarUrl",
          "updatedByEmail",
          "updatedByName"
        ],
        "type": "object"
      },
...
      "SchemaDto": {
        "properties": {
          "archivedAt": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "organizationId": {
            "format": "int64",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "archivedAt",
          "createdAt",
          "description",
          "id",
          "name",
          "organizationId",
          "updatedAt"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MLSuite API",
    "version": "contract"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/analyzer/schema": {
      "post": {
        "operationId": "generateSchema",
        "parameters": [
          {
            "in": "query",
            "name": "oneHotSeparator",
            "required": false,
            "schema": {
              "default": "__",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "dataframe": {
                    "format": "binary",
                    "type": "string"
                  },
                  "model": {
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "model"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": { },
                  "type": "object"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "analyzer-controller"
        ]
      }
    },
...
    "/api/models/{modelId}/archive": {
      "post": {
        "operationId": "archive_1",
        "parameters": [
          {
            "in": "path",
            "name": "modelId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "model-controller"
        ]
      }
    },
    "/api/models/{modelId}/duplicate": {
      "post": {
        "operationId": "duplicate_2",
        "parameters": [
          {
            "in": "path",
            "name": "modelId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "model-controller"
        ]
      }
    },
...
    "/api/organizations/{organizationId}/roles/{roleId}/duplicate": {
      "post": {
        "operationId": "duplicate_1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "roleId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DuplicateRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleDefinitionDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "role-catalog-controller"
        ]
      }
    },
...
    "/api/schema-bookmarks/{bookmarkId}": {
      "get": {
        "operationId": "get_1",
        "parameters": [
          {
            "in": "path",
            "name": "bookmarkId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaBookmarkDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-bookmark-controller"
        ]
      },
      "put": {
        "operationId": "move",
        "parameters": [
          {
            "in": "path",
            "name": "bookmarkId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveSchemaBookmarkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaBookmarkDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-bookmark-controller"
        ]
      }
    },
...
    "/api/schemas": {
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 24,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "updated",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "active",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageDtoSchemaCatalogItemDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-controller"
        ]
      },
      "post": {
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSchemaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-controller"
        ]
      }
    },
...
    "/api/schemas/{schemaId}": {
      "delete": {
        "operationId": "delete_1",
        "parameters": [
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "schema-controller"
        ]
      },
      "get": {
        "operationId": "get_2",
        "parameters": [
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-controller"
        ]
      },
      "patch": {
        "operationId": "rename",
        "parameters": [
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-controller"
        ]
      }
    },
    "/api/schemas/{schemaId}/archive": {
      "post": {
        "operationId": "archive",
        "parameters": [
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-controller"
        ]
      }
    },
    "/api/schemas/{schemaId}/bookmarks": {
      "get": {
        "operationId": "list_3",
        "parameters": [
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SchemaBookmarkDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-bookmark-controller"
        ]
      },
      "post": {
        "operationId": "create_3",
        "parameters": [
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSchemaBookmarkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaBookmarkDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-bookmark-controller"
        ]
      }
    },
...
    "/api/schemas/{schemaId}/duplicate": {
      "post": {
        "operationId": "duplicate",
        "parameters": [
          {
            "in": "path",
            "name": "schemaId",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "versionId",
            "required": false,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDto"
                }
              }
            },
            "description": "OK"
          }
        },
        "tags": [
          "schema-controller"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}