# api/src/main/java/dev/ulloasp/mlsuite/schema/adapter/in/web/SchemaBookmarkController.java
package dev.ulloasp.mlsuite.schema.adapter.in.web;

import java.util.List;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import dev.ulloasp.mlsuite.schema.application.dto.CreateSchemaBookmarkRequest;
import dev.ulloasp.mlsuite.schema.application.dto.MoveSchemaBookmarkRequest;
import dev.ulloasp.mlsuite.schema.application.dto.PredictBookmarkDto;
import dev.ulloasp.mlsuite.schema.application.dto.SchemaBookmarkDto;
import dev.ulloasp.mlsuite.schema.application.port.in.SchemaBookmarkUseCase;
import dev.ulloasp.mlsuite.schema.application.port.in.PredictBookmarkCatalogUseCase;
import dev.ulloasp.mlsuite.security.identity.CurrentUser;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;

@RestController
@RequestMapping("/api")
@RequiredArgsConstructor
public class SchemaBookmarkController {

    private final SchemaBookmarkUseCase bookmarkUseCase;
    private final PredictBookmarkCatalogUseCase predictCatalog;

    @GetMapping("/schemas/{schemaId}/bookmarks")
    public ResponseEntity<List<SchemaBookmarkDto>> list(CurrentUser user,
            @PathVariable Long schemaId) {
        return ResponseEntity.ok(bookmarkUseCase.listBookmarks(user.userId(), schemaId).stream()
                .map(SchemaBookmarkDto::from)
                .toList());
    }

    @GetMapping("/schema-bookmarks")
    public ResponseEntity<List<PredictBookmarkDto>> listOrganization(CurrentUser user) {
        return ResponseEntity.ok(predictCatalog.listBookmarks(user.userId()));
    }

    @PostMapping("/schemas/{schemaId}/bookmarks")
    public ResponseEntity<SchemaBookmarkDto> create(CurrentUser user, @PathVariable Long schemaId,
            @Valid @RequestBody CreateSchemaBookmarkRequest request) {
        return ResponseEntity.status(HttpStatus.CREATED).body(SchemaBookmarkDto.from(
                bookmarkUseCase.createBookmark(user.userId(), schemaId, request)));
    }

    @GetMapping("/schema-bookmarks/{bookmarkId}")
    public ResponseEntity<SchemaBookmarkDto> get(CurrentUser user, @PathVariable Long bookmarkId) {
        return ResponseEntity.ok(SchemaBookmarkDto.from(
                bookmarkUseCase.getBookmark(user.userId(), bookmarkId)));
    }

    @PutMapping("/schema-bookmarks/{bookmarkId}")
    public ResponseEntity<SchemaBookmarkDto> move(CurrentUser user, @PathVariable Long bookmarkId,
            @Valid @RequestBody MoveSchemaBookmarkRequest request) {
        return ResponseEntity.ok(SchemaBookmarkDto.from(
                bookmarkUseCase.moveBookmark(user.userId(), bookmarkId, request)));
    }
}